[TypeDeclaration] Extract shared array return resolution into StrictReturnNewArrayResolver service
